refactor(ladybug): split store.rs monolith into store/{init,helpers,query,path,traversal,mutate} (CR-001)
The 2113-line ladybug_store/store.rs (3.7x the 500-line Rust limit, a
readability blocker flagged by codesize + readability auditors) is split into a
store/ submodule along its existing responsibility seams, each file with a
context header and under the limit:
  mod.rs       router + GraphStore trait impl
  helpers.rs   constants, types, value extractors, lbug_call choke-point
  init.rs      struct, constructor, FFI primitives, schema DDL
  traversal.rs private BFS helpers
  path.rs      find_path + query_cross_boundary
  query.rs     read-path methods
  mutate.rs    write-path methods

Pure structural refactor: no method body, Cypher, signature, or error message
changed. Test count unchanged (3852 passed / 25 ignored), clippy clean.

Refs #146
